Moved auth http code to HttpClientExtensions
