+++ FirebaseAdmin/FirebaseAdmin/Auth/FirebaseUserManager.cs
@@ -39,8 +39,6 @@ namespace FirebaseAdmin.Auth
         private readonly ConfigurableHttpClient httpClient;
         private readonly string baseUrl;
 
-        private readonly AuthErrorHandler errorHandler;
-
         internal FirebaseUserManager(FirebaseUserManagerArgs args)
         {
             if (string.IsNullOrEmpty(args.ProjectId))
@@ -51,7 +49,6 @@ namespace FirebaseAdmin.Auth
 
             this.httpClient = args.ClientFactory.CreateAuthorizedHttpClient(args.Credential);
             this.baseUrl = string.Format(IdTooklitUrl, args.ProjectId);
-            this.errorHandler = new AuthErrorHandler();
         }
 
         public void Dispose()
@@ -69,27 +66,6 @@ namespace FirebaseAdmin.Auth
             };
 
             return new FirebaseUserManager(args);
-        }
-
-        internal static FirebaseAuthException HandleHttpError(HttpRequestException exception)
-        {
-            var temp = exception.ToFirebaseException();
-            return new FirebaseAuthException(
-                temp.ErrorCode,
-                temp.Message,
-                inner: temp.InnerException,
-                response: temp.HttpResponse);
-        }
-
-        internal static FirebaseAuthException HandleParseError(
-            Exception e, HttpResponseMessage response)
-        {
-            throw new FirebaseAuthException(
-                ErrorCode.Unknown,
-                "Error while parsing Auth service response.",
-                AuthErrorCode.UnexpectedResponse,
-                e,
-                response);
         }
 
         /// <summary>
@@ -277,15 +253,15 @@ namespace FirebaseAdmin.Auth
             return new UserRecord(result.Users[0]);
         }
 
-        private async Task<Extensions.ParsedResponseInfo<TResult>> PostAndDeserializeAsync<TResult>(
+        private async Task<HttpExtensions.ParsedResponseInfo<TResult>> PostAndDeserializeAsync<TResult>(
             string path, object body, CancellationToken cancellationToken)
         {
             var response = await this.PostAsync(path, body, cancellationToken)
                 .ConfigureAwait(false);
-            return response.SafeDeserialize<TResult>(HandleParseError);
+            return response.SafeDeserialize<TResult>();
         }
 
-        private async Task<Extensions.ResponseInfo> PostAsync(
+        private async Task<HttpExtensions.ResponseInfo> PostAsync(
             string path, object body, CancellationToken cancellationToken)
         {
             var request = new HttpRequestMessage()
@@ -297,13 +273,12 @@ namespace FirebaseAdmin.Auth
             return await this.SendAsync(request, cancellationToken).ConfigureAwait(false);
         }
 
-        private async Task<Extensions.ResponseInfo> SendAsync(
+        private async Task<HttpExtensions.ResponseInfo> SendAsync(
             HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            var response = await this.httpClient.SendAndReadAsync(
-                request, cancellationToken, HandleHttpError);
-            this.errorHandler.ThrowIfError(response.HttpResponse, response.Body);
-            return response;
+            return await this.httpClient
+                .SendAndReadAsync(request, cancellationToken)
+                .ConfigureAwait(false);
         }
 
         /// <summary>
